numero,contacto,usuario,rol

+++ src/main/java/upc/com/visiontech2/serviceinterfaces/IRolService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface IRolService {
     public List<Role> list();
     public void insert(Role r);
-    public Role listId(int idRol);
+    public Role listId(long idRol);
     public void update(Role r);
-    public void delete(int idRol);
+    public void delete (long idRol);
 }
